Создает модуль data.js
